refactor(worktree): split WorktreeDetailRefactored.tsx into focused modules (#755)
#728〜#747 で 6 回連続改修により 2205 行に肥大化していた WorktreeDetailRefactored.tsx を、責務別に分割し保守性を改善。受入基準 <1000 行を達成 (2205→610 行、-1595)。

主な変更:
- src/components/worktree/WorktreeDetailRefactored.tsx: 2205→610 行
  - orchestrator (薄い分岐) に縮小、PC/Mobile dispatch のみ
- src/components/worktree/WorktreeDetailDesktop.tsx (新規, 696 行):
  - PC Desktop レイアウト制御 (ActivityBar / TerminalContainer / renderSplitPane)
- src/components/worktree/WorktreeDetailMobile.tsx (新規, 368 行):
  - Mobile タブ制御 + MobileContent を SubComponents から抽出
- src/components/worktree/WorktreeDetailSubComponents.tsx: 1128→804 行
  - MobileContent / MobileInfoContent を WorktreeDetailMobile.tsx へ移管
- src/hooks/useWorktreeDetailController.ts (新規, 1489 行):
  - 親コンポーネントの body logic (state/effect/handler) を抽出
- src/hooks/usePendingInsertText.ts (新規, 116 行):
  - TODO:[D1-001] 解消: pendingInsertTextMap (Map<splitIndex, string>) 管理を hook 化
- tests/unit/hooks/usePendingInsertText.test.tsx (新規, 10 tests pass)
- src/hooks/useFileOperations.ts: MoveTarget 型を export 追加
- tests/integration/issue-278-acceptance.test.ts: pre-existing 2 FAIL を解消 (8/10→10/10)
- CLAUDE.md: モジュール表更新

検証 (全 PASS):
- npx tsc --noEmit ✅
- npm run lint ✅ No warnings or errors
- npm run test:unit ✅ 365 files / 6788 passed / 7 skipped
- npm run build:server ✅
- npm run build (Next.js) ✅

データ契約保持:
- testid: terminal-split-container / add-terminal-split / terminal-split-pane-* /
  split-resizer-* / mobile-tab-* / split-status-indicator-* / desktop-agent-status-* など
- 既存ユニットテスト 全 PASS = 振る舞い無回帰
